Add a test-oriented attribute for functions startups.
Fixes #154.

An alternative approach would be to allow the existing
FunctionsStartupAttribute to be applied to types as well as the
assembly... in which case we'd probably want to change the framework
to find startup classes associated with the function itself. That
could potentially be quite useful...

Very happy to discuss those options.