Update all examples to use FunctionsStartupAttribute specified on the function type
This is more convenient than doing it in an assembly attribute, partly because an assembly attribute can't be within a namespace declaration, so needs a fully-qualified type name.

The docs do include an example of how you would specify it at the assembly level though.

Additionally, this commit changes the documentation around specifying startup classes in tests.