Example of 'Reading' from Firestore
NOTE: if I am mistaken the Google.Cloud.Firestore dependency needs to be installed in the local machine so that it is available in the deployment.  I have provided instruction in the attached as to how I configured the build.  

First MANY thanks for your efforts, I do not mean my comments or observations to come across as negative, I am just trying to be succinct.

You may feel free to remove what instruction you feel is redundant, as I say there are a LOT of ways to get this wrong, and mostly the deployment fails silently or there is just no response.  This had be on the verge of reverting back to Azure, just to have more mature dotnet workflows.  A big part of the problem (took me to the breaking point and I really cannot understand what exactly is going on is the "Type Tuples" [FirestoreData] & [FirestoreProperty] They aren't types and they aren't tuples but they APPEAR to be Required for the "ConvertTo()" to correctly destructure the DocumentSnapshot.  

 Please be aware that there are MANY ways for this to fail and when testing with a deployed function you will get precious little feedback
    If you are having trouble; pleas check that... 
    1)    The Firestore DB must be defined and populated CORRECTLY.
    2)    The deploy is correctly configured (verify the project-ids between Firestore DB and Function are aligned.
    3)    The payload in the function test aligns with the structure of the Firestore document