chore: Refactor release-please to be a single component
We release all of the Functions Framework together, even though it's
four packages - like GAX. We don't want or need to keep track of the
versions or history for the separate packages separately.

Note that previously we *have* created separate release tags for the
separate packages. I believe we'll no longer do this after this
change, but that shouldn't affect releasing.