chore: Change how a missing CloudEventAdapter is handled
In .NET 6.0, using a null-returning provider would require the
null-forgiving operator, which would be unfortunate (and could lead
to problems down the road, if the framework performs earlier
validation).

Instead, we introduce a new-and-immediately-obsolete constructor
which doesn't require a formatter. The DI framework will invoke this
if there is no formatter available, and it just throws an exception
with a clear message.